Moved get_processed_source_to_response_media_type to separate module to allow re-use
